refactor(input): move toolbar event application
