+++ src/input/state/core/toolbar/apply/pages.rs
