+++ src/input/state/core/toolbar/mod.rs
@@ -1,0 +1,3 @@
+//! Input-side toolbar event application.
+
+mod apply;
